Allow 2D tensors in Binary Binned AUPRC (but not 3D or more)
Summary:
I mistakenly disallowed NxM tensors in the function `_binary_binned_auprc_update_input_check`. This made me unable to make a class-version of the `BinaryBinnedAUPRC`, as calling `update` would result in an error. Therefore, I remove this.

Also, I disallow 3D (or more) tensors for Binned AUROC (which was originally not there).

Reviewed By: bobakfb, ananthsub

Differential Revision: D44522159

fbshipit-source-id: 6bfc2777f459f1a25b3fb6d7630c676789da6d45