support `rank` argument in sync_states (#181)
Summary:
Pull Request resolved: https://github.com/pytorch/torcheval/pull/181

# Context
TorchEval used to support syncing states on specific ranks. When rewriting the sync logic, we removed this feature to maintain simplicity. We want to add this feature back for checkpointing flexibility purposes

# This Diff
Added support for `rank` argument for methods in synclib. Next (and final) diff will add support for `rank` argument in toolkit (so for methods like `sync_and_compute`, etc that act on the metric object)

Also refactored few lines here and there for readability

Reviewed By: galrotem

Differential Revision: D49039113

fbshipit-source-id: 303821da2681edc495598d6a1107ddab1b7ee1c1