Fix warning in aggregation.mean (#187)
Summary:
Pull Request resolved: https://github.com/pytorch/torcheval/pull/187

This diff fixes the incorrect warning when running `mean.compute()` when the mean is exactly 0.

Instead of checking for the weighted sum of elements to be 0, we instead check for the total sum of weights to be zero (meaning that the average can be 0 without error, but we throw a warning when dividing by zero)

We also update the error message to reflect that the issue is no weight has been accumulated, since it is possible to call this function with only 0 weights.

Addresses: https://github.com/pytorch/torcheval/issues/185

Reviewed By: JKSenthil

Differential Revision: D50806243

fbshipit-source-id: 04d75826ae8c1a24cc3718967d86bdd982081538