Bug fix: recall macro average computation when not all classes are present in target and input
Summary: Both f1 and precision are able to perform macro averaging when the input and target both do not contained instances of a class (e.g if num_classes = 5 and both input/target contain only instances of 4 classes). However, recall computation fails in such a case. This diff adds a missing line to make this case work.

Reviewed By: JKSenthil

Differential Revision: D62046138

fbshipit-source-id: 3d6706ed22b8143d61edf8b4891dec6f378b0723